fix: split utilities and add dictionary cog
transfered dictionary commands from utilities to disctionary cog
